plant changes with button press

+++ florish/lib/globals.dart
@@ -3,4 +3,4 @@ library globals;
 int waterCount = 0;
 int drinkCount = 0;
 
-int bac = 0;
+double bac = 0.00;
